Provide a is_remote method on mount
Due to windows associating remote with a disk and unix with mount info.
This method needs to be provided on mount.

Unix uses heuristics to determine whethere a mount is remove while
windows associates it with a disk. As such this method requires access
to either disk or mount info which is only possible on mount.